requirements.txt
numpy==1.21.6

tqdm
typeguard
visualdl>=2.2.0
opencv-python <= 4.6.0
PyYAML
shapely
scipy
terminaltables
Cython
pycocotools
setuptools

# for vehicleplate
pyclipper

# for mot
lap
motmetrics
sklearn==0.0
filterpy

# flask
Flask>=2.1.0

paddlepaddle-gpu==2.4.1
